# back/docs/swagger.yaml
openapi: 3.0.3
info:
  title: ManageFlow API
  version: "1.0.0"
  description: |
    API do **ManageFlow** — Sistema de filas de espera para restaurantes (TCC).
    - Autenticação via **JWT Bearer** na maioria das rotas.
    - Rotas **Públicas** não exigem token.
servers:
  - url: "{baseUrl}/api"
    description: Ambiente configurado por variável
    variables:
      baseUrl:
        default: http://localhost:3001
        description: URL base (use PUBLIC_API_BASE_URL em produção)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdEmpresaPath:
      name: idEmpresa
      in: path
      required: true
      schema: { type: integer }
      description: ID da empresa
    IdUsuarioPath:
      name: idUsuario
      in: path
      required: true
      schema: { type: integer }
...
    IdFilaPath:
      name: id_fila
      in: path
      required: true
      schema: { type: integer }
...
    DtMovtoPath:
      name: dtMovto
      in: path
      required: true
      schema:
        type: string
        format: date
      description: Data do movimento (YYYY-MM-DD)
  schemas:
...
    Empresa:
      type: object
      properties:
        ID_EMPRESA: { type: integer, example: 10 }
        NOME_EMPRESA: { type: string, example: "Restaurante Exemplo" }
        ENDERECO: { type: string, example: "Av. Brasil, 100" }
        CIDADE: { type: string, example: "Curitiba" }
        UF: { type: string, example: "PR" }
        CEP: { type: string, example: "80000-000" }
        IMG_LOGO: { type: string, example: "/uploads/logo/10.png" }
        IMG_LOGO_URL: { type: string, example: "https://api.dominio.com/uploads/logo/10.png" }
    PerfilEmpresa:
      type: object
      properties:
        ID_PERFIL: { type: integer, example: 1 }
        NOME_PERFIL: { type: string, example: "ADM" }
        NIVEL:
          type: integer
          description: 9=ADM; 5=STAF; 1=ANALYST (exemplo)
...
    ConfiguracaoFila:
      type: object
      properties:
        ID_CONF_FILA: { type: integer, example: 55 }
        ID_EMPRESA: { type: integer, example: 10 }
        NOME_FILA: { type: string, example: "Fila Principal" }
        INI_VIG: { type: string, format: date, example: "2025-10-01" }
        FIM_VIG: { type: string, format: date, nullable: true }
        SITUACAO: { type: integer, example: 1 }
        TOKEN_FILA: { type: string, example: "d122ed7b-35c9-41be-afd7-7b3d42" }
        PER_LOC: { type: integer, example: 1 }
        TOLERANCIA_MIN: { type: integer, example: 10 }
        IMG_LOGO: { type: string, nullable: true }
        IMG_BANNER: { type: string, nullable: true }
        IMG_LOGO_URL: { type: string, nullable: true }
        IMG_BANNER_URL: { type: string, nullable: true }
...
    ClienteFila:
      type: object
      properties:
        ID_CLIENTE_FILA: { type: integer, example: 1001 }
        ID_FILA: { type: integer, example: 777 }
        NOME: { type: string, example: "Beatriz Sakai" }
        CPF: { type: string, nullable: true }
        TELEFONE: { type: string, nullable: true }
        QTD_PESSOAS: { type: integer, nullable: true }
        STATUS:
          type: string
          enum: [AGUARDANDO, CHAMADO, CONFIRMADO, NAO_COMPARECEU, ATENDIDO, REMOVIDO]
          example: AGUARDANDO
        DT_ENTRA: { type: string, format: date-time }
...
    AtualizarSituacaoClienteInput:
      type: object
      required: [situacao]
      properties:
        situacao:
          type: string
          enum: [AGUARDANDO, CHAMADO, CONFIRMADO, NAO_COMPARECEU, ATENDIDO, REMOVIDO]
          example: CHAMADO
    EnviarNotificacaoInput:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
          example: chamada
...
security:
  - bearerAuth: []
paths:
...
  /empresas-do-usuario/{idUsuario}:
    get:
      tags: [Empresas]
      summary: Empresas às quais o usuário tem acesso
      parameters: [ { $ref: "#/components/parameters/IdUsuarioPath" } ]
      responses:
        "200":
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Empresa" }
  /detalhes/{idEmpresa}:
    get:
      tags: [Empresas]
      summary: Detalhes da empresa
      parameters: [ { $ref: "#/components/parameters/IdEmpresaPath" } ]
      responses:
        "200":
          description: Empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Empresa" }
  /perfis/{idEmpresa}:
    get:
      tags: [Usuários & Permissões]
      summary: Perfis (RBAC) da empresa
      parameters: [ { $ref: "#/components/parameters/IdEmpresaPath" } ]
      responses:
        "200":
          description: Perfis disponíveis
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PerfilEmpresa" }
...
  /filas/{idEmpresa}:
    get:
      tags: [Configuração de Fila]
      summary: Configurações da empresa específica
      parameters: [ { $ref: "#/components/parameters/IdEmpresaPath" } ]
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConfiguracaoFila" }

  # ======== PÚBLICO (TOKEN_FILA) ========
...
  /fila/{idEmpresa}/{dtMovto}/{id_fila}/clientes:
    get:
      tags: [Fila do Dia]
      summary: Listar clientes da fila do dia
      parameters:
        - $ref: "#/components/parameters/IdEmpresaPath"
        - $ref: "#/components/parameters/DtMovtoPath"
        - $ref: "#/components/parameters/IdFilaPath"
      responses:
        "200":
          description: Clientes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClienteFila" }
  /fila/{idEmpresa}/{dtMovto}/{id_fila}/adicionar-cliente:
    post:
      tags: [Fila do Dia]
      summary: Inserir cliente manualmente
      parameters:
        - $ref: "#/components/parameters/IdEmpresaPath"
        - $ref: "#/components/parameters/DtMovtoPath"
        - $ref: "#/components/parameters/IdFilaPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string }
                cpf: { type: string, nullable: true }
                telefone: { type: string, nullable: true }
                qtdPessoas: { type: integer, nullable: true }
                observacao: { type: string, nullable: true }
      responses:
        "201":
          description: Cliente inserido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteFila" }
  /fila/{idEmpresa}/{dtMovto}/{id_fila}/cliente/{idCliente}/atualizar-situacao:
    put:
      tags: [Fila do Dia]
      summary: Atualizar situação do cliente (emite eventos)
      parameters:
        - $ref: "#/components/parameters/IdEmpresaPath"
        - $ref: "#/components/parameters/DtMovtoPath"
        - $ref: "#/components/parameters/IdFilaPath"
        - name: idCliente
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AtualizarSituacaoClienteInput" }
      responses:
        "200": { description: Situação atualizada }
  /fila/{idEmpresa}/{dtMovto}/{id_fila}/cliente/{idCliente}/enviar-notificacao:
    post:
      tags: [Fila do Dia]
      summary: Enviar notificação (e agendar ausência quando aplicável)
      parameters:
        - $ref: "#/components/parameters/IdEmpresaPath"
        - $ref: "#/components/parameters/DtMovtoPath"
        - $ref: "#/components/parameters/IdFilaPath"
        - name: idCliente
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnviarNotificacaoInput" }
      responses:
        "200": { description: Notificação enviada }

  # ======== DASHBOARD ========
...